appease the hound

+++ app/controllers/memberships_controller.rb
@@ -1,11 +1,9 @@
+# frozen_string_literal: true
+
 class MembershipsController < ApplicationController
   def destroy
     membership = Membership.find(params[:id])
-
-    unless current_user == membership.user
-      membership.destroy
-    end
-
+    membership.destroy unless current_user == membership.user
     redirect_to edit_group_path membership.group
   end
 end
